requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
lxml==4.9.1
multidict==6.0.2
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.10
yarl==1.7.2
